jiayue-cloud-functions/temp/explore_phila_data/requirements.txt
protobuf==6.30.2
pyproj==3.6.1
python-dotenv==1.1.0
Requests==2.32.3
Shapely==2.0.7
